Fix improper handling of extended inode which spans blocks
An extended inode may consume the last 32 bytes of one block and the
next 32 bytes of another. The inodes can be loaded independently of
where they land on the datablock. This issue is fixed by directly using
the bytes buffer rather than loading inodes into its position in the
block buffer.

Signed-off-by: Derek McGowan <derek@mcg.dev>